backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
mediapipe==0.10.5
numpy==1.26.2
scipy==1.11.4
python-jose==3.3.0
pyyaml==6.0.1
google-generativeai==0.3.2
Pillow==10.1.0
